fix conflicts with users and migrations
