+++ Back/src/app.ts
@@ -22,9 +22,9 @@ export const startServer = async () => {
 
 	const host = '0.0.0.0';
 
-	app.register(fastifyJwt, {
-	  secret: process.env.FASTIFY_SECRET,
-	});
+	// app.register(fastifyJwt, {
+	//   secret: process.env.FASTIFY_SECRET,
+	// });
 
 	app.register(registerRoutes);
